# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ChangeProof AI API",
    "description": "API de traçabilité et documentation automatisée par IA. SaaS B2B multi-tenant.",
    "version": "0.1.0",
    "contact": {
      "name": "ChangeProof AI",
      "url": "https://changeproof.ai"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API locale (Next.js App Router)"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/changes": {
      "get": {
        "tags": ["Changes"],
        "summary": "Lister les changements",
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "source", "in": "query", "schema": { "type": "string", "enum": ["MANUAL", "JIRA", "GITHUB", "GITLAB", "CONFLUENCE"] } },
          { "name": "project", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["PENDING", "RUNNING", "COMPLETED", "FAILED", "NEEDS_REVIEW", "VALIDATED", "REJECTED"] } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100, "maximum": 200 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": {
            "description": "Liste paginée des changements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "changes": { "type": "array", "items": { "$ref": "#/components/schemas/Change" } },
                    "total": { "type": "integer" },
                    "offset": { "type": "integer" },
                    "limit": { "type": "integer" }
                  }
                }
              }
            }
          },
          "401": { "description": "Non autorisé" },
          "500": { "description": "Erreur serveur" }
        }
      },
      "post": {
        "tags": ["Changes"],
        "summary": "Créer un changement",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "source", "projectId"],
                "properties": {
                  "title": { "type": "string", "minLength": 1, "maxLength": 500, "example": "Mise à jour du module d'authentification" },
                  "description": { "type": "string", "maxLength": 10000 },
                  "source": { "type": "string", "enum": ["MANUAL", "JIRA", "GITHUB", "GITLAB", "CONFLUENCE"] },
                  "projectId": { "type": "string" },
                  "rawContent": { "type": "string", "maxLength": 50000 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Changement créé", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Change" } } } },
          "400": { "description": "Champs requis manquants ou source invalide" },
          "401": { "description": "Non autorisé" },
          "404": { "description": "Projet introuvable" },
          "500": { "description": "Erreur serveur" }
        }
      }
    },
    "/changes/{id}": {
      "get": {
        "tags": ["Changes"],
        "summary": "Obtenir un changement",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Détails du changement", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Change" } } } },
          "404": { "description": "Changement introuvable" }
        }
      },
      "patch": {
        "tags": ["Changes"],
        "summary": "Mettre à jour un changement",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Changement mis à jour" },
          "404": { "description": "Changement introuvable" }
        }
      },
      "delete": {
        "tags": ["Changes"],
        "summary": "Supprimer un changement",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Changement supprimé" },
          "404": { "description": "Changement introuvable" }
        }
      }
    },
...
    "/ai/analyze": {
      "post": {
        "tags": ["AI"],
        "summary": "Analyser un changement avec l'IA",
        "description": "Lance une analyse IA structurée sur un changement (ticket JIRA, PR, diff git, etc.). L'IA retourne un résumé, les impacts, les risques et les documents à mettre à jour.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "jiraTicket": { "type": "string", "description": "ID ou description du ticket JIRA" },
                  "functionalDescription": { "type": "string", "description": "Description fonctionnelle du changement" },
                  "pullRequest": { "type": "string", "description": "URL ou contenu de la Pull Request" },
                  "gitDiff": { "type": "string", "description": "Diff git du changement" },
                  "modifiedFiles": { "type": "string", "description": "Liste des fichiers modifiés" },
                  "oldDocumentation": { "type": "string", "description": "Ancienne documentation de référence" },
                  "newDocumentation": { "type": "string", "description": "Nouvelle documentation proposée" },
                  "developerComments": { "type": "string", "description": "Commentaires du développeur" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Analyse IA réussie",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AIStructuredResult" }
              }
            }
          },
          "400": { "description": "Aucun champ rempli ou contenu bloqué par le filtre de sécurité" },
          "401": { "description": "Non autorisé" },
          "500": { "description": "Erreur lors de l'analyse IA" }
        }
      }
    },
    "/ai/provider": {
      "get": {
        "tags": ["AI"],
        "summary": "Lister les providers IA",
        "description": "Retourne tous les providers IA configurés pour l'organisation (sans les clés API).",
        "responses": {
          "200": {
            "description": "Liste des providers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/AIProvider" } }
                  }
                }
              }
            }
          },
          "401": { "description": "Non autorisé" }
        }
      },
      "post": {
        "tags": ["AI"],
        "summary": "Ajouter un provider IA",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AIProviderCreate" }
            }
          }
        },
        "responses": {
          "201": { "description": "Provider créé", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/AIProvider" } } } } } },
          "400": { "description": "Validation échouée" },
          "401": { "description": "Non autorisé" },
          "500": { "description": "Erreur serveur" }
        }
      },
      "patch": {
        "tags": ["AI"],
        "summary": "Mettre à jour un provider IA",
        "parameters": [
          { "name": "testConnection", "in": "query", "schema": { "type": "boolean" }, "description": "Si true, teste la connexion au lieu de mettre à jour" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AIProviderUpdate" }
            }
          }
        },
        "responses": {
          "200": { "description": "Provider mis à jour" },
          "400": { "description": "Validation échouée" },
          "401": { "description": "Non autorisé" },
          "404": { "description": "Provider introuvable" }
        }
      },
      "delete": {
        "tags": ["AI"],
        "summary": "Supprimer un provider IA",
        "parameters": [
          { "name": "id", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Provider supprimé" },
          "400": { "description": "ID requis" },
          "401": { "description": "Non autorisé" },
          "404": { "description": "Provider introuvable" }
        }
      }
    },
    "/search": {
      "get": {
        "tags": ["Search"],
        "summary": "Recherche globale",
        "description": "Recherche dans les titres et descriptions des changements, documents et tickets.",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string", "minLength": 1, "maxLength": 500 }, "description": "Terme de recherche" }
        ],
        "responses": {
          "200": {
            "description": "Résultats de recherche",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "result": { "$ref": "#/components/schemas/SearchResult" }
                  }
                }
              }
            }
          },
          "400": { "description": "Paramètre 'q' requis" },
          "401": { "description": "Non autorisé" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "Cookie de session Next-Auth"
      }
    },
    "schemas": {
...
      "Change": {
        "type": "object",
        "required": ["id", "source", "title", "projectId", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "string" },
          "source": { "type": "string", "enum": ["MANUAL", "JIRA", "GITHUB", "GITLAB", "CONFLUENCE"] },
          "title": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "rawContent": { "type": "string", "nullable": true },
          "projectId": { "type": "string" },
          "project": { "type": "object", "properties": { "id": { "type": "string" }, "name": { "type": "string" } } },
          "createdBy": { "type": "object", "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "email": { "type": "string" } } },
          "analyses": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "status": { "type": "string" },
                "confidence": { "type": "number" }
              }
            }
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "AIProvider": {
        "type": "object",
        "required": ["id", "name", "type", "baseUrl", "defaultModel", "isActive", "createdAt"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "type": { "type": "string", "example": "openai_compatible" },
          "baseUrl": { "type": "string", "format": "uri" },
          "defaultModel": { "type": "string", "example": "mistral-medium" },
          "embeddingModel": { "type": "string", "nullable": true },
          "timeout": { "type": "integer" },
          "maxTokens": { "type": "integer" },
          "temperature": { "type": "number" },
          "streaming": { "type": "boolean" },
          "jsonMode": { "type": "boolean" },
          "toolCalling": { "type": "boolean" },
          "isActive": { "type": "boolean" },
          "lastTestAt": { "type": "string", "format": "date-time", "nullable": true },
          "lastTestSuccess": { "type": "boolean", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "AIProviderCreate": {
        "type": "object",
        "required": ["name", "baseUrl", "apiKey", "defaultModel"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 200 },
          "type": { "type": "string", "default": "openai_compatible" },
          "baseUrl": { "type": "string", "format": "uri" },
          "apiKey": { "type": "string", "minLength": 10, "maxLength": 512, "description": "Clé API (chiffrée côté serveur)" },
          "defaultModel": { "type": "string", "minLength": 1, "maxLength": 100 },
          "embeddingModel": { "type": "string", "maxLength": 100 },
          "timeout": { "type": "integer", "minimum": 1000, "maximum": 300000, "default": 60000 },
          "maxTokens": { "type": "integer", "minimum": 100, "maximum": 128000, "default": 4096 },
          "temperature": { "type": "number", "minimum": 0, "maximum": 2, "default": 0.3 },
          "streaming": { "type": "boolean", "default": false },
          "jsonMode": { "type": "boolean", "default": true },
          "toolCalling": { "type": "boolean", "default": false },
          "isActive": { "type": "boolean", "default": true }
        }
      },
      "AIProviderUpdate": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": { "type": "string", "description": "ID du provider à mettre à jour" },
          "name": { "type": "string" },
          "type": { "type": "string" },
          "baseUrl": { "type": "string", "format": "uri" },
          "apiKey": { "type": "string", "description": "Nouvelle clé API (optionnelle)" },
          "defaultModel": { "type": "string" },
          "embeddingModel": { "type": "string" },
          "timeout": { "type": "integer" },
          "maxTokens": { "type": "integer" },
          "temperature": { "type": "number" },
          "streaming": { "type": "boolean" },
          "jsonMode": { "type": "boolean" },
          "toolCalling": { "type": "boolean" },
          "isActive": { "type": "boolean" }
        }
      },
      "AIStructuredResult": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "globalSummary": { "type": "string", "description": "Résumé global du changement" },
          "businessSummary": { "type": "string", "description": "Résumé métier" },
          "technicalSummary": { "type": "string", "description": "Résumé technique" },
          "impactedComponents": { "type": "array", "items": { "type": "string" } },
          "impactedApis": { "type": "array", "items": { "type": "string" } },
          "impactedScreens": { "type": "array", "items": { "type": "string" } },
          "impactedUserRoles": { "type": "array", "items": { "type": "string" } },
          "impactedData": { "type": "array", "items": { "type": "string" } },
          "impactedConfig": { "type": "array", "items": { "type": "string" } },
          "impactedSecurity": { "type": "array", "items": { "type": "string" } },
          "externalDependencies": { "type": "array", "items": { "type": "string" } },
          "functionalRisk": { "type": "object", "properties": { "level": { "type": "string" }, "description": { "type": "string" } } },
          "technicalRisk": { "type": "object", "properties": { "level": { "type": "string" }, "description": { "type": "string" } } },
          "securityRisk": { "type": "object", "properties": { "level": { "type": "string" }, "description": { "type": "string" } } },
          "operationalRisk": { "type": "object", "properties": { "level": { "type": "string" }, "description": { "type": "string" } } },
          "confidenceLevel": { "type": "number" },
          "confirmed": { "type": "array", "items": { "type": "string" } },
          "probable": { "type": "array", "items": { "type": "string" } },
          "unproven": { "type": "array", "items": { "type": "string" } },
          "missingInfo": { "type": "array", "items": { "type": "string" } },
          "questionsToAsk": { "type": "array", "items": { "type": "string" } },
          "documentsToUpdate": { "type": "array", "items": { "type": "string" } },
          "finalRecommendation": { "type": "string" }
        }
      },
      "SearchResult": {
        "type": "object",
        "properties": {
          "summary": { "type": "string" },
          "sources": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": { "type": "string" },
                "id": { "type": "string" },
                "title": { "type": "string" }
              }
            }
          },
          "tickets": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": { "id": { "type": "string" }, "title": { "type": "string" } }
            }
          },
          "commits": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": { "id": { "type": "string" }, "message": { "type": "string" } }
            }
          },
          "documents": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": { "id": { "type": "string" }, "title": { "type": "string" } }
            }
          },
          "confidence": { "type": "number" }
        }
      }
    }
  }
}